+++ python/ngrok/utils.py
@@ -1,8 +1,7 @@
 import typing as t
 from contextlib import contextmanager
 import ngrok
-import functools
-import asyncio
+
 
 
 @contextmanager
